# openapi.yaml
openapi: 3.1.0
info:
  title: Pipeline Labs API
  version: 2.0.0
  description: |
    Pipeline Labs is an SDK-first platform for orchestrating LLM and ML training workflows
    on infrastructure owned by the customer. The API supports Supabase bearer tokens for the web app
    and server-issued API keys for SDK and automation use cases.
servers:
  - url: https://pipelinelabs-ashen.vercel.app/api
    description: Production
  - url: http://localhost:3001/api
    description: Local development
...
paths:
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Resolve current identity
      operationId: getCurrentIdentity
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Authenticated identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentIdentity'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/api-keys:
    get:
      tags: [Auth]
      summary: List API keys
      operationId: listApiKeys
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: API keys for the current user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Auth]
      summary: Create API key
      operationId: createApiKey
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API key created. The plaintext key is only returned once.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiKey'
                  - type: object
                    properties:
                      key:
                        type: string
                        example: plk_live_7d8b6f7fd4a8b8c7d2e4f0bc24901ef6f51ef984d4be6cf1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/export:
    get:
      tags: [Auth]
      summary: Export all user-scoped SDK data
      operationId: exportUserData
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Full export of the authenticated user's control-plane data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDataExport'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/api-keys/{id}:
    delete:
      tags: [Auth]
      summary: Revoke API key
      operationId: revokeApiKey
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: API key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /cloud/providers:
    get:
      tags: [Cloud]
      summary: List connected cloud providers
      operationId: listCloudProviders
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Connected cloud accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CloudProvider'
    post:
      tags: [Cloud]
      summary: Add cloud provider
      operationId: createCloudProvider
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudProviderInput'
      responses:
        '200':
          description: Connected provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudProvider'
        '400':
          $ref: '#/components/responses/BadRequest'
  /cloud/providers/{id}:
    patch:
      tags: [Cloud]
      summary: Update cloud provider
      operationId: updateCloudProvider
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                credentials:
                  type: object
                  additionalProperties: true
                is_active:
                  type: boolean
      responses:
        '200':
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudProvider'
    delete:
      tags: [Cloud]
      summary: Delete cloud provider
      operationId: deleteCloudProvider
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Provider deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /cloud/estimate-cost:
    post:
      tags: [Cloud]
      summary: Estimate training cost
      operationId: estimateTrainingCost
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                providerId:
                  type: string
                  format: uuid
                gpuType:
                  type: string
                  enum: [none, T4, A10G, V100, A100]
                estimatedHours:
                  type: number
                  default: 1
              required: [providerId, gpuType]
      responses:
        '200':
          description: Estimated cost
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    type: string
                  gpuType:
                    type: string
                  estimatedHours:
                    type: number
                  estimatedCostUsd:
                    type: number
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token used by the web app.
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Server-issued API key for SDK and automation workflows.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CurrentIdentity:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        authType:
          type: string
          enum: [jwt, api_key, dev]
        apiKeyId:
          type: string
          format: uuid
          nullable: true
        profile:
          oneOf:
            - $ref: '#/components/schemas/Profile'
            - type: 'null'
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
        last4:
          type: string
        scopes:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
    CreateApiKeyRequest:
      type: object
      properties:
        name:
          type: string
        expiresAt:
          type: string
          format: date-time
        scopes:
          type: array
          items:
            type: string
      required: [name]
...
    DatasetAnalysis:
      type: object
      properties:
        data_type:
          type: string
          enum: [tabular, text, image, audio, mixed]
        row_count:
          type: integer
        column_count:
          type: integer
        columns:
          type: array
          items:
            $ref: '#/components/schemas/DatasetColumn'
        label_column:
          type: string
          nullable: true
        task_type:
          type: string
        suggested_models:
          type: array
          items:
            $ref: '#/components/schemas/SuggestedModel'
        preprocessing_steps:
          type: array
          items:
            type: string
        data_quality_score:
          type: number
    DatasetColumn:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [numerical, categorical, text, datetime, id]
        unique_count:
          type: integer
        null_percent:
          type: number
        sample_values:
          type: array
          items:
            type: string
    SuggestedModel:
      type: object
      properties:
        name:
          type: string
        framework:
          type: string
        gpu_required:
          type: string
        estimated_time_minutes:
          type: number
        estimated_cost_usd:
          type: number
        confidence:
          type: number
    TrainingPlan:
      type: object
      properties:
        task:
          type: string
        model:
          type: string
        framework:
          type: string
          enum: [pytorch, tensorflow, sklearn, xgboost]
        train_test_split:
          type: string
        batch_size:
          type: integer
        optimizer:
          type: string
        learning_rate:
          type: number
        epochs:
          type: integer
        gpu_required:
          type: string
          enum: [none, T4, A10G, V100, A100]
        preprocessing:
          type: array
          items:
            type: string
        metrics:
          type: array
          items:
            type: string
        estimated_time_minutes:
          type: number
        estimated_cost_usd:
          type: number
        hyperparameter_search:
          type: object
          nullable: true
        reasoning:
          type: string
          nullable: true
    TrainingPlanRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        dataset_id:
          type: string
          format: uuid
        plan:
          $ref: '#/components/schemas/TrainingPlan'
        dataset_analysis:
          $ref: '#/components/schemas/DatasetAnalysis'
        status:
          type: string
          enum: [draft, approved, rejected, executing, completed, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DatasetRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        file_name:
          type: string
        mime_type:
          type: string
          nullable: true
        storage_path:
          type: string
          nullable: true
        row_count:
          type: integer
        column_count:
          type: integer
        status:
          type: string
        data_type:
          type: string
          nullable: true
        label_column:
          type: string
          nullable: true
        task_type:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CloudProviderInput:
      type: object
      properties:
        provider:
          type: string
          enum: [aws, azure, gcp, runpod]
        label:
          type: string
        credentials:
          type: object
          additionalProperties: true
      required: [provider, credentials]
    CloudProvider:
      allOf:
        - $ref: '#/components/schemas/CloudProviderInput'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            is_active:
              type: boolean
            last_verified_at:
              type: string
              format: date-time
              nullable: true
            created_at:
              type: string
              format: date-time
    TrainingJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        dataset_id:
          type: string
          format: uuid
        cloud_provider_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [pending, provisioning, installing, running, completed, failed, cancelled]
        instance_id:
          type: string
          nullable: true
        instance_type:
          type: string
          nullable: true
        instance_ip:
          type: string
          nullable: true
        instance_region:
          type: string
          nullable: true
        final_metrics:
          type: object
          additionalProperties:
            type: number
        model_artifact_path:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TrainingJobExport:
      allOf:
        - $ref: '#/components/schemas/TrainingJob'
        - type: object
          properties:
            plan_id:
              type: string
              format: uuid
            dataset_id:
              type: string
              format: uuid
            cloud_provider_id:
              type: string
              format: uuid
              nullable: true
...
    UserDataExport:
      type: object
      properties:
        exportedAt:
          type: string
          format: date-time
        userId:
          type: string
          format: uuid
        authType:
          type: string
          enum: [jwt, api_key, dev]
        profile:
          oneOf:
            - $ref: '#/components/schemas/Profile'
            - type: 'null'
        datasets:
          type: array
          items:
            $ref: '#/components/schemas/DatasetRecord'
        cloudProviders:
          type: array
          items:
            $ref: '#/components/schemas/CloudProvider'
        trainingPlans:
          type: array
          items:
            $ref: '#/components/schemas/TrainingPlanRecord'
        trainingJobs:
          type: array
          items:
            $ref: '#/components/schemas/TrainingJobExport'
        apiKeys:
          type: array
          items:
            $ref: '#/components/schemas/ApiKey'
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
...